+++ CHANGELOG.md
@@ -8,6 +8,12 @@
   to properly accept wild cards for Win32_NetworkAdapter filters.
 - Updated MSFT_Netbios to use 'Format-Win32NetworkADapterFilterByNetConnectionID'
   - fixes [Issue #413](https://github.com/PowerShell/NetworkingDsc/issues/413).
+- Corrected minor style and consistency issues in `NetworkingDsc.Common.tests.ps1`
+  and `NetworkingDsc.Common.ps1`.
+- Changed verbose messages in `Test-DscParameterState` to include
+  full type name.
+- Fixed bug in `Test-DscParameterState` that causes it to return true when
+  both the current array and desired array is empty.
 
 ## 7.3.0.0